Add function for default 16x16 grid
Reworked the script to generate a default 16x16 canvas. Clicking the
button prompts the user for a new canvas size. The program will repeat
the prompt until the user types in a number 1 - 100.
All the required features have been implemented for this practice
project. The extra features might be added later.